Implement Gnomes split for Any%, 100%, & Long Path
The Any% and Long Path routes are handled by checking if Alexander has
managed to get into the swamp or the bookworm's room from the beach. The
100% route is handled by checking for the 2 points earned by fooling the
final gnome. These points are always awarded at a very specific X,Y
location that is unique from all the other points you can earn in this
room.