Make the cliff base conditional on the gnomes
In the 100% run, you need to visit the cliff base to get the flower
before being able to fool the gnomes. But you don't normally climb the
cliffs until you have everything you need to finish the catacombs, such
as the red scarf. Therefore, this split only makes sense to do when you
visit the cliff base after having fooled or bypassed the gnomes.