Change Score data type from byte to short
Based on experiments, the game treats the score as a short rather than
as a byte, even though its valid range is from 0 to 231. If you edit
this area of memory, you can get the game to display a score as high as
999 before weird glitches begin to occur. This is very useful for us
because it makes it easier to detect an invalid score, which can help us
detect whether we are in the "Initial" state or the "Restart" state.