Distinguish the "Initial" and "Restart" states
WHen the user clicks the "Restart" button inside the game, all the
offsets for the game data in memory are shifted by a few bytes. This
only happens once: any subsequent restarts preserve the memory
locations. If the user saves a game while in this mode, the save game
file remembers it. The user can also return to "Initial" state by
restoring a save that was made during that state.

Although it would be reasonable to ask users to avoid the Restart button
(since there is a better way to reset a speedrun in this game) it
makes sense to dynamically detect which state we are in at runtime.