Move performSplit() functions from startup to init
In the startup action we can only write to the settings, not read from
them. This limitation also affects Func objects defined inside the
startup action. By moving the definition of the performSplit functions
from the startup action to the init action, we gain access to the
readable version of the settings object, removing the need for us to
cache the options in a dictionary on every update.